Refine Jake chat personality
