+++ lib/jake-chat/message-markdown.ts
@@ -1,0 +1,8 @@
+import type { UIMessage } from "ai"
+
+export function getMessageMarkdown(message: UIMessage) {
+  return message.parts
+    .map((part) => (part.type === "text" ? part.text : ""))
+    .join("")
+    .trim()
+}
